backend/requirements.production.txt
# FastAPI y servidor web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
alembic==1.12.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==45.0.4

# Validación y serialización
pydantic==2.4.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Manejo de archivos y formularios
python-multipart==0.0.6
aiofiles==23.2.1

# Procesamiento de documentos
Pillow==10.4.0
PyMuPDF==1.24.11
docx2pdf==0.1.8

# Notificaciones push
pywebpush==1.14.0

# Programación de tareas
APScheduler==3.10.1

# Utilidades
python-dotenv==1.0.0
pytz==2023.3
requests==2.32.4
httpx==0.28.1

# Análisis de datos
pandas==2.0.3
numpy==1.24.3 